requirements.txt
pandas>=1.5.0
numpy>=1.23.0
jieba>=0.42.1
openpyxl>=3.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
google-generativeai>=0.3.0
streamlit>=1.28.0
plotly>=5.17.0
pyyaml>=6.0
scikit-learn>=1.3.0
tqdm>=4.65.0